+++ src/components/Tooltip/types.ts
@@ -1,0 +1,11 @@
+export type TooltipPlacement =
+  | "top"
+  | "topRight"
+  | "topLeft"
+  | "bottom"
+  | "bottomRight"
+  | "bottomLeft"
+  | "right"
+  | "left";
+
+export type TooltipPosition = [number, number];