+++ src/components/Tooltip/useTooltip.ts
@@ -1,0 +1,9 @@
+import * as React from "react";
+
+import TooltipContext from "./TooltipContext";
+
+const useTooltip = () => {
+  return React.useContext(TooltipContext);
+};
+
+export default useTooltip;