+++ src/data/data.ts
@@ -204,7 +204,7 @@ export const data = {
       ]
     },
     {
-      name: "app/scr/component/util/myUtil/number2/me.tsx",
+      name: "app/scr/component/util/myUtil/number2/verylongone/longer/more/me.tsx",
       children: [
         { name: "Arrays", value: 8258 },
         { name: "Colors", value: 10001 },